requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
collection==0.1.6
distro==1.9.0
exceptiongroup==1.2.2
gpg==1.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
joblib==1.4.2
libcomps==0.1.18
mypy==1.11.2
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.0.2
openai==1.44.1
pandas==2.2.2
pycode-similar==1.4
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rpm==4.16.1.3
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sklearn==0.0
sniffio==1.3.1
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.5
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
